pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "adaptiverag"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "beautifulsoup4>=4.12",
    "pypdf>=4.0",
    "python-docx>=1.1",
    "numpy>=1.26",
    "chromadb>=0.4",
    "faiss-cpu>=1.7",
    "sentence-transformers>=2.2",
    "openai>=1.0",
]

[tool.setuptools.packages.find]
where = ["src"]